替换到 DNMVCS 最新版

+++ src/SwooleSessionHandler.php
@@ -2,8 +2,9 @@
 namespace DNMVCS\SwooleHttpd;
 
 use DNMVCS\SwooleHttpd\SwooleSingleton;
+use SessionHandlerInterface;
 
-class SwooleSessionHandler implements \SessionHandlerInterface
+class SwooleSessionHandler implements SessionHandlerInterface
 {
     use SwooleSingleton;
     private $savePath;
@@ -38,7 +39,11 @@ class SwooleSessionHandler implements \SessionHandlerInterface
     }
     public function gc($maxlifetime)
     {
-        foreach (glob("$this->savePath/sess_*") as $file) {
+        $files=glob("$this->savePath/sess_*");
+        if (!$files) {
+            return true;
+        }
+        foreach ($files as $file) {
             if (filemtime($file) + $maxlifetime < time() && file_exists($file)) {
                 unlink($file);
             }
